Compliance test fixes
