+++ .gitignore
@@ -10,6 +10,7 @@ spec/rp_env/
 .vagrant/
 .bundle/
 Gemfile.lock
+sec_results/
 vendor/
 junit/
 log/
