+++ spec/acceptance/suites/compliance/00_test_prep_spec.rb
@@ -1,0 +1,1 @@
+../default/00_test_prep_spec.rb
