+++ spec/acceptance/suites/compliance/05_basic_test_spec.rb
@@ -1,0 +1,1 @@
+../default/05_basic_test_spec.rb
