+++ spec/acceptance/suites/compliance/include
@@ -1,0 +1,1 @@
+../default/include
