--- spec/acceptance/suites/default/nodesets
@@ -1,1 +1,0 @@
-../../nodesets
